delete要求時に削除キーを照合してDBから削除する
