+++ src/main/java/models/posts/PostContribution.java
@@ -5,12 +5,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import databases.entities.Contribution;
 import models.contributions.HandleContribution;
+import models.payloads.PostPayload;
 import spark.Request;
 import spark.Response;
 
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class PostContribution extends Status {
     private HandleDB handleDB = new HandleDB();
@@ -26,7 +25,7 @@ public class PostContribution extends Status {
 
         try {
             // postPayloadを生成する
-            PostPayload payload = new ObjectMapper().readValue(unescapeUnicode(request.body()), PostPayload.class);
+            PostPayload payload = new ObjectMapper().readValue(handleContribution.unescapeUnicode(request.body()), PostPayload.class);
             if (!payload.isValid()) {
                 return setBadRequest(response);
             }
@@ -52,37 +51,6 @@ public class PostContribution extends Status {
             // ステータスコード400 BadRequestを設定する
             return setBadRequest(response);
         }
-    }
-
-    /**
-     * unicodeエスケープされた文字列を元に戻す
-     * @param body unicodeが含まれる文字列
-     * @return アンエスケープした文字列
-     */
-    private String unescapeUnicode(String body) {
-        String regex = "\\\\\\\\u([a-fA-F0-9]{4})";
-        Pattern p = Pattern.compile(regex);
-        Matcher m = p.matcher(body);
-
-        StringBuffer strBuf = new StringBuffer();
-        while (m.find()) {
-            if (m.group(1).matches("000a|000d"))
-                m.appendReplacement(strBuf, "\\\\\\\\n"); // 改行、ラインフィールドは改行コードとして保存する
-            else
-                m.appendReplacement(strBuf, Matcher.quoteReplacement(String.valueOf((char) Integer.parseInt(m.group(1), 16))));
-        }
-        m.appendTail(strBuf);
-
-        return validateBody(strBuf.toString());
-    }
-
-    /**
-     * タグを削除した文字列を返す
-     * @param body タグが含まれる文字列
-     * @return タグを削除した文字列
-     */
-    private String validateBody(String body) {
-        return body.replaceAll("<(\".*?\"|'.*?'|[^'\"])*?>", "");
     }
 
     /**
