+++ src/main/resources/META-INF/databases/daos/ContributionDao/deleteById.sql
@@ -3,3 +3,5 @@ FROM
     contribution
 WHERE
     id = /* id */0
+    AND
+    delete_key = /* deleteKey */'hoge'
