+++ src/test/java/models/posts/PostContributionTest.java
@@ -84,32 +84,4 @@ public class PostContributionTest {
         // verify
         verify(response).status(200);
     }
-
-    @Test
-    public void unicodeエスケープされた文字列を元に戻す() throws Exception {
-        // setup
-        String escapeStr = "{\"username\":\"\\\\u897f\\\\u6728\\\\u91ce\\\\u771f\\\\u59eb\", \"content\":\"\\\\u000a\"}";
-        Method method = postContribution.getClass().getDeclaredMethod("unescapeUnicode", String.class);
-        method.setAccessible(true);
-
-        // exercise
-        String unescapeStr = (String) method.invoke(postContribution, escapeStr);
-
-        // verify
-        assertThat(unescapeStr, is("{\"username\":\"西木野真姫\", \"content\":\"\\\\n\"}"));
-    }
-
-    @Test
-    public void HTMLタグを除いた文字列を返す() throws Exception {
-        // setup
-        String str = "{\"content\":\"hoge<b>foo</b>bar\"}";
-        Method method = postContribution.getClass().getDeclaredMethod("validateBody", String.class);
-        method.setAccessible(true);
-
-        // exercise
-        String excludedStr = (String) method.invoke(postContribution, str);
-
-        // verify
-        assertThat(excludedStr, is("{\"content\":\"hogefoobar\"}"));
-    }
 }
